# openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "EventHubAPI | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/api/admin/imports": {
      "post": {
        "tags": [
          "AdminImports"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventImportRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventImportRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventImportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportStartResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportStartResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportStartResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "AdminImports"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventImportRun"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventImportRun"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EventImportRun"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/imports/{importId}": {
      "get": {
        "tags": [
          "AdminImports"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportRun"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportRun"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportRun"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/imports/{importId}/items/{itemId}": {
      "get": {
        "tags": [
          "AdminImports"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportItem"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminImports"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EventImportEditDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/EventImportEditDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/EventImportEditDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportItem"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportItem"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportItem"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/imports/{importId}/items/{itemId}/confirm": {
      "post": {
        "tags": [
          "AdminImports"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportConfirmResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportConfirmResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventImportConfirmResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/imports/{importId}/items/{itemId}/retry": {
      "post": {
        "tags": [
          "AdminImports"
        ],
        "parameters": [
          {
            "name": "importId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "itemId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Accepted"
          }
        }
      }
    },
    "/api/admin/tags": {
      "post": {
        "tags": [
          "AdminTags"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/tags/{tagId}": {
      "put": {
        "tags": [
          "AdminTags"
        ],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminTags"
        ],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/admin/images": {
      "post": {
        "tags": [
          "Images"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "file"
                ],
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StoredImage"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredImage"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredImage"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/media/{objectKey}": {
      "get": {
        "tags": [
          "Images"
        ],
        "parameters": [
          {
            "name": "objectKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "200": {
            "description": "Изображение",
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/png": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/webp": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              },
              "image/gif": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/tags/{tagId}": {
      "get": {
        "tags": [
          "Tags"
        ],
        "parameters": [
          {
            "name": "tagId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagResponse"
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EventImportConfirmResponse": {
        "required": [
          "eventId"
        ],
        "type": "object",
        "properties": {
          "eventId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "EventImportEditDto": {
        "required": [
          "title",
          "description",
          "eventDateTime",
          "location",
          "deadline",
          "tagIds"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventDateTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "mainImg": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "EventImportItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "importRunId": {
            "type": "string",
            "format": "uuid"
          },
          "sourceKey": {
            "type": "string"
          },
          "source": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EventImportStatus"
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "description": {
            "type": [
              "null",
              "string"
            ]
          },
          "originalDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventDateTime": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "deadline": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "location": {
            "type": [
              "null",
              "string"
            ]
          },
          "mainImg": {
            "type": [
              "null",
              "string"
            ]
          },
          "tagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "suggestedTagIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "suggestedDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "warnings": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": [
              "null",
              "string"
            ]
          },
          "eventId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "leaseUntil": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "leaseToken": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          }
        }
      },
      "EventImportRequest": {
        "required": [
          "urls"
        ],
        "type": "object",
        "properties": {
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "EventImportRun": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EventImportItem"
            }
          }
        }
      },
      "EventImportStartResponse": {
        "required": [
          "importId"
        ],
        "type": "object",
        "properties": {
          "importId": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "EventImportStatus": {
        "type": "integer"
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "StoredImage": {
        "required": [
          "objectKey",
          "url"
        ],
        "type": "object",
        "properties": {
          "objectKey": {
            "type": "string"
          },
          "url": {
            "type": "string"
          }
        }
      },
...
      "TagRequest": {
        "required": [
          "name",
          "description",
          "examples"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "TagResponse": {
        "required": [
          "id",
          "name",
          "description",
          "examples"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "examples": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": []
    }
  ],
...
}